requirements.txt
# requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.24.0
websockets==12.0
openai==1.3.7
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpcore==0.17.3
httpx==0.24.1

